附件上传和下载 个人资料填写

+++ Src/Caviar.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -90,6 +90,8 @@ namespace Caviar.Infrastructure.Persistence
         public virtual async Task<T> UpdateEntityAsync<T>(T entity, bool isSaveChange = true) where T : class, IUseEntity,new()
         {
             IsEntityNull(entity);
+            var dbEntity = await SingleOrDefaultAsync<T>(u => u.Id == entity.Id);
+            if (dbEntity == null) throw new ArgumentException("非法操作，修改未授权数据");
             DbContext.Entry(entity).State = EntityState.Modified;
             if (isSaveChange)
             {
@@ -108,6 +110,8 @@ namespace Caviar.Infrastructure.Persistence
         public virtual async Task<T> UpdateEntityAsync<T>(T entity, Expression<Func<T, object>> fieldExp, bool isSaveChange = true) where T : class, IUseEntity,new()
         {
             IsEntityNull(entity);
+            var dbEntity = await SingleOrDefaultAsync<T>(u => u.Id == entity.Id);
+            if (dbEntity == null) throw new ArgumentException("非法操作，修改未授权数据");
             DbContext.Entry(entity).Property(fieldExp).IsModified = true;
             if (isSaveChange)
             {
@@ -127,6 +131,8 @@ namespace Caviar.Infrastructure.Persistence
         public virtual async Task<int> UpdateEntityAsync<T>(IEnumerable<T> entity, bool isSaveChange = true) where T : class, IUseEntity,new()
         {
             IsEntityNull(entity);
+            var dbEntity = GetEntityAsync<T>(u => entity.Contains(u));
+            if (dbEntity.Count() != entity.Count()) throw new ArgumentException("非法操作，修改未授权数据");
             DbContext.UpdateRange(entity);
             if (isSaveChange)
             {
@@ -147,28 +153,14 @@ namespace Caviar.Infrastructure.Persistence
         public virtual async Task<bool> DeleteEntityAsync<T>(T entity, bool isSaveChange = true, bool IsDelete = false) where T : class, IUseEntity, new()
         {
             IsEntityNull(entity);
-            //逻辑删除在后面更新
-            DbContext.Entry(entity).State = EntityState.Deleted;
+            var dbEntity = await SingleOrDefaultAsync<T>(u=>u.Id == entity.Id);
+            if (dbEntity == null) throw new ArgumentException("非法操作，删除未授权数据");
+            DbContext.Entry(dbEntity).State = EntityState.Deleted;
             if (isSaveChange)
             {
                 await SaveChangesAsync();
             }
             return true;
-            //if (entity.IsDelete || IsDelete)
-            //{
-            //    DbContext.Entry(entity).State = EntityState.Deleted;
-            //    if (isSaveChange)
-            //    {
-            //        await SaveChangesAsync();
-            //    }
-            //    return true;
-            //}
-            //else
-            //{
-            //    entity.IsDelete = true;
-            //    await UpdateEntityAsync(entity, isSaveChange);
-            //    return false;
-            //}
         }
         /// <summary>
         /// 批量删除
@@ -181,6 +173,8 @@ namespace Caviar.Infrastructure.Persistence
         public virtual async Task<int> DeleteEntityAsync<T>(IEnumerable<T> entity, bool isSaveChange = true, bool IsDelete = false) where T : class, IUseEntity, new()
         {
             IsEntityNull(entity);
+            var dbEntity = GetEntityAsync<T>(u => entity.Contains(u));
+            if (dbEntity.Count() != entity.Count()) throw new ArgumentException("非法操作，删除未授权数据");
             var removeList = entity.Where(u => u.IsDelete).ToList();//取出物理删除数据
             DbContext.RemoveRange(removeList);
             removeList = entity.Where(u => u.IsDelete == false).ToList();//取出逻辑删除数据
